+++ app/include/utils/misc.hpp
@@ -2,6 +2,7 @@
 
 #include <string>
 #include <cstdint>
+#include <vector>
 
 namespace misc {
 
@@ -10,6 +11,8 @@ std::string sec2Time(int64_t t);
 std::string randHex(const int len);
 
 std::string hexEncode(const unsigned char* data, size_t len);
+
+std::vector<std::string> split(const std::string& data, char seq);
 
 bool sendIPC(const std::string& sock, const std::string& payload);