+++ app/src/view/mpv_core.cpp
@@ -523,6 +523,7 @@ void MPVCore::eventMainLoop() {
 
 void MPVCore::reset() {
     brls::Logger::debug("MPVCore::reset");
+    mpvCoreEvent.fire(MpvEventEnum::RESET);
     this->video_stopped = true;
     this->duration = 0;     // second
     this->cache_speed = 0;  // Bps